`ircevents` → `msgtimes`; `irclog` → `ircevents`
